[so/testing] rename s/testutils/testdata/g directory to make happy kitchen CI tests
