+++ pkg/network/so/testdata/procfs/1/maps
