+++ pkg/network/so/testdata/procfs/1/task/1/maps
