+++ pkg/network/so/testdata/procfs/138104/task/138104/maps
