+++ pkg/network/so/testdata/procfs/138105/task/138110/maps
