+++ pkg/network/so/testdata/procfs/370680/task/370680/maps
